Fix all tests but xcebra tests
